+++ lib/kiwi_codec/decode_error.ex
@@ -1,3 +1,7 @@
 defmodule KiwiCodec.DecodeError do
+  @moduledoc """
+  Raised when Kiwi wire data or binary schema data cannot be decoded.
+  """
+
   defexception [:message]
 end