+++ lib/kiwi_codec/encode_error.ex
@@ -1,3 +1,7 @@
 defmodule KiwiCodec.EncodeError do
+  @moduledoc """
+  Raised when a value cannot be encoded as Kiwi wire data.
+  """
+
   defexception [:message]
 end