+++ lib/kiwi_codec/message_props.ex
@@ -1,5 +1,7 @@
 defmodule KiwiCodec.MessageProps do
-  @moduledoc false
+  @moduledoc """
+  Compiled Kiwi definition metadata for generated modules.
+  """
 
   alias KiwiCodec.FieldProps